teardown_automation_project-teardown_automation 2/requirements.txt
crewai==0.148.0
flask>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
openai>=1.13.3
python-dotenv>=1.0.0